+++ crates/hyperswitch_domain_models/src/payments/payment_attempt.rs
@@ -777,6 +777,7 @@ pub struct PaymentAttempt {
     /// Whether external 3DS authentication was attempted for this payment.
     /// This is based on the configuration of the merchant in the business profile
     pub external_three_ds_authentication_attempted: Option<bool>,
+    pub external_threeds_authentication_type: Option<common_enums::DecoupledAuthenticationType>,
     /// The connector that was used for external authentication
     pub authentication_connector: Option<String>,
     /// The foreign key reference to the authentication details
@@ -964,6 +965,7 @@ impl PaymentAttempt {
             encoded_data: None,
             merchant_connector_id: None,
             external_three_ds_authentication_attempted: None,
+            external_threeds_authentication_type: None,
             authentication_connector: None,
             authentication_id: None,
             fingerprint_id: None,
@@ -1057,6 +1059,7 @@ impl PaymentAttempt {
             encoded_data: None,
             merchant_connector_id: Some(request.merchant_connector_id.clone()),
             external_three_ds_authentication_attempted: None,
+            external_threeds_authentication_type: None,
             authentication_connector: None,
             authentication_id: None,
             fingerprint_id: None,
@@ -1158,6 +1161,7 @@ impl PaymentAttempt {
             encoded_data: None,
             merchant_connector_id: None,
             external_three_ds_authentication_attempted: None,
+            external_threeds_authentication_type: None,
             authentication_connector: None,
             authentication_id: None,
             fingerprint_id: None,
@@ -1282,6 +1286,7 @@ impl PaymentAttempt {
             encoded_data: None,
             merchant_connector_id: request.payment_merchant_connector_id.clone(),
             external_three_ds_authentication_attempted: None,
+            external_threeds_authentication_type: None,
             authentication_connector: None,
             authentication_id: None,
             fingerprint_id: None,
@@ -1386,6 +1391,7 @@ pub struct PaymentAttempt {
     pub unified_code: Option<String>,
     pub unified_message: Option<String>,
     pub external_three_ds_authentication_attempted: Option<bool>,
+    pub external_threeds_authentication_type: Option<common_enums::DecoupledAuthenticationType>,
     pub authentication_connector: Option<String>,
     pub authentication_id: Option<id_type::AuthenticationId>,
     pub mandate_data: Option<MandateDetails>,
@@ -1981,6 +1987,7 @@ pub enum PaymentAttemptUpdate {
         updated_by: String,
         merchant_connector_id: Option<id_type::MerchantConnectorAccountId>,
         external_three_ds_authentication_attempted: Option<bool>,
+        external_threeds_authentication_type: Option<common_enums::DecoupledAuthenticationType>,
         authentication_connector: Option<String>,
         authentication_id: Option<id_type::AuthenticationId>,
         payment_method_billing_address_id: Option<String>,
@@ -2146,6 +2153,7 @@ pub enum PaymentAttemptUpdate {
     AuthenticationUpdate {
         status: storage_enums::AttemptStatus,
         external_three_ds_authentication_attempted: Option<bool>,
+        external_threeds_authentication_type: Option<common_enums::DecoupledAuthenticationType>,
         authentication_connector: Option<String>,
         authentication_id: Option<id_type::AuthenticationId>,
         updated_by: String,
@@ -2308,6 +2316,7 @@ impl PaymentAttemptUpdate {
                 merchant_connector_id: connector_id,
                 payment_method_id,
                 external_three_ds_authentication_attempted,
+                external_threeds_authentication_type,
                 authentication_connector,
                 authentication_id,
                 payment_method_billing_address_id,
@@ -2348,6 +2357,7 @@ impl PaymentAttemptUpdate {
                 merchant_connector_id: connector_id,
                 payment_method_id,
                 external_three_ds_authentication_attempted,
+                external_threeds_authentication_type,
                 authentication_connector,
                 authentication_id,
                 payment_method_billing_address_id,
@@ -2697,12 +2707,14 @@ impl PaymentAttemptUpdate {
             Self::AuthenticationUpdate {
                 status,
                 external_three_ds_authentication_attempted,
+                external_threeds_authentication_type,
                 authentication_connector,
                 authentication_id,
                 updated_by,
             } => DieselPaymentAttemptUpdate::AuthenticationUpdate {
                 status,
                 external_three_ds_authentication_attempted,
+                external_threeds_authentication_type,
                 authentication_connector,
                 authentication_id,
                 updated_by,
@@ -2944,6 +2956,7 @@ impl behaviour::Conversion for PaymentAttempt {
             net_amount: Some(self.net_amount.get_total_amount()),
             external_three_ds_authentication_attempted: self
                 .external_three_ds_authentication_attempted,
+            external_threeds_authentication_type: self.external_threeds_authentication_type,
             authentication_connector: self.authentication_connector,
             authentication_id: self.authentication_id,
             mandate_data: self.mandate_data.map(Into::into),
@@ -3080,6 +3093,8 @@ impl behaviour::Conversion for PaymentAttempt {
                 unified_message: storage_model.unified_message,
                 external_three_ds_authentication_attempted: storage_model
                     .external_three_ds_authentication_attempted,
+                external_threeds_authentication_type: storage_model
+                    .external_threeds_authentication_type,
                 authentication_connector: storage_model.authentication_connector,
                 authentication_id: storage_model.authentication_id,
                 mandate_data: storage_model.mandate_data.map(Into::into),
@@ -3191,6 +3206,7 @@ impl behaviour::Conversion for PaymentAttempt {
             net_amount: Some(self.net_amount.get_total_amount()),
             external_three_ds_authentication_attempted: self
                 .external_three_ds_authentication_attempted,
+            external_threeds_authentication_type: self.external_threeds_authentication_type,
             authentication_connector: self.authentication_connector,
             authentication_id: self.authentication_id,
             mandate_data: self.mandate_data.map(Into::into),
@@ -3276,6 +3292,7 @@ impl behaviour::Conversion for PaymentAttempt {
             encoded_data,
             merchant_connector_id,
             external_three_ds_authentication_attempted,
+            external_threeds_authentication_type,
             authentication_connector,
             authentication_id,
             fingerprint_id,
@@ -3362,6 +3379,7 @@ impl behaviour::Conversion for PaymentAttempt {
                 .and_then(|details| details.unified_message.clone()),
             net_amount,
             external_three_ds_authentication_attempted,
+            external_threeds_authentication_type,
             authentication_connector,
             authentication_id,
             fingerprint_id,
@@ -3517,6 +3535,8 @@ impl behaviour::Conversion for PaymentAttempt {
                 merchant_connector_id: storage_model.merchant_connector_id,
                 external_three_ds_authentication_attempted: storage_model
                     .external_three_ds_authentication_attempted,
+                external_threeds_authentication_type: storage_model
+                    .external_threeds_authentication_type,
                 authentication_connector: storage_model.authentication_connector,
                 authentication_id: storage_model.authentication_id,
                 fingerprint_id: storage_model.fingerprint_id,
@@ -3581,6 +3601,7 @@ impl behaviour::Conversion for PaymentAttempt {
             encoded_data,
             merchant_connector_id,
             external_three_ds_authentication_attempted,
+            external_threeds_authentication_type,
             authentication_connector,
             authentication_id,
             fingerprint_id,
@@ -3664,6 +3685,7 @@ impl behaviour::Conversion for PaymentAttempt {
                 .and_then(|details| details.unified_message.clone()),
             net_amount: amount_details.net_amount,
             external_three_ds_authentication_attempted,
+            external_threeds_authentication_type,
             authentication_connector,
             authentication_id,
             fingerprint_id,